+++ build.rs
@@ -1,7 +1,6 @@
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     tonic_build::configure()
-        .build_transport(false)
-        .out_dir("src/google")
+        .build_server(false)
         .compile(
             &["googleapis/google/cloud/bigquery/storage/v1/storage.proto"],
             &["googleapis"],